# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_huypham67_bookmark-service_internal_dto_auth.LoginRequest:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_huypham67_bookmark-service_internal_dto_auth.LoginResponse:
    properties:
      data:
        type: string
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_pkg_response.Pagination'
    type: object
  github_com_huypham67_bookmark-service_internal_dto_auth.RegisterUserRequest:
    properties:
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  github_com_huypham67_bookmark-service_internal_dto_auth.RegisterUserResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_auth.UserData'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_pkg_response.Pagination'
    type: object
  github_com_huypham67_bookmark-service_internal_dto_auth.UserData:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkData:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkData'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_pkg_response.Pagination'
    type: object
  github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkData'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_pkg_response.Pagination'
    type: object
  github_com_huypham67_bookmark-service_internal_dto_bookmark.CreateBookmarkRequest:
    properties:
      description:
        maxLength: 500
        minLength: 2
        type: string
      url:
        type: string
    required:
    - description
    - url
    type: object
  github_com_huypham67_bookmark-service_internal_dto_health.HealthCheckResponse:
    properties:
      instance_id:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_link.ShortenURLRequest:
    properties:
      exp:
        maximum: 86400
        minimum: 0
        type: integer
      url:
        maxLength: 2048
        type: string
    required:
    - url
    type: object
  github_com_huypham67_bookmark-service_internal_dto_link.ShortenURLResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_profile.UpdateUserRequest:
    properties:
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      email:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_profile.UpdateUserResponse:
    type: object
  github_com_huypham67_bookmark-service_internal_dto_profile.UserData:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  github_com_huypham67_bookmark-service_internal_dto_profile.UserResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_profile.UserData'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_huypham67_bookmark-service_pkg_response.Pagination'
    type: object
  github_com_huypham67_bookmark-service_pkg_response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: This is the API documentation for the Bookmark Service
  title: Bookmark Service API
  version: "1.1"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      description: Check application health status and Redis connection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_health.HealthCheckResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_health.HealthCheckResponse'
      summary: Health Check
      tags:
      - health
  /v1/bookmarks:
    get:
      consumes:
      - application/json
      description: Get paginated list of bookmarks for the authenticated user
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: 'Items per page, max 100 (default: 10)'
        in: query
        name: limit
        type: integer
      - default: created_at
        description: 'Sort field: created_at, updated_at, code, url (default: created_at)'
        enum:
        - created_at
        - updated_at
        - code
        - url
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of bookmarks with pagination
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: List Bookmarks
      tags:
      - bookmarks
    post:
      consumes:
      - application/json
      description: Create a new bookmark for the authenticated user
      parameters:
      - description: Bookmark creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_bookmark.CreateBookmarkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bookmark created successfully
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_bookmark.BookmarkResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Bookmark code already exists
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Create Bookmark
      tags:
      - bookmarks
  /v1/bookmarks/{id}:
    delete:
      description: Delete an existing bookmark
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bookmark deleted successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid bookmark ID
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bookmark not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Delete Bookmark
      tags:
      - bookmarks
    put:
      consumes:
      - application/json
      description: Update an existing bookmark (description and/or URL)
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated bookmark data
        in: body
        name: request
        schema:
          properties:
            description:
              type: string
            url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Bookmark updated successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid bookmark ID or request data
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bookmark not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Update Bookmark
      tags:
      - bookmarks
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      description: 'Redirect user to the original URL for a given code. The code''s
        routing prefix selects the backing store: shortened links resolve from Redis,
        bookmarks resolve from the database.'
      parameters:
      - description: 'Short code (prefixed: a-h for links, i-z for bookmarks)'
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect successful
        "404":
          description: Short link not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Redirect to Original URL
      tags:
      - links
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Create a shortened URL code and save it to Redis
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_link.ShortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Shorten URL generated successfully
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_link.ShortenURLResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Shorten URL
      tags:
      - links
  /v1/self/info:
    get:
      consumes:
      - application/json
      description: Get authenticated user information from JWT token
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_profile.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Get User Info
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update authenticated user's display name and email
      parameters:
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_profile.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_profile.UpdateUserResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Update User Info
      tags:
      - users
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Login a user with username and password
      parameters:
      - description: User login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_auth.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Login User
      tags:
      - users
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, username, and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_auth.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/github_com_huypham67_bookmark-service_internal_dto_auth.RegisterUserResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Register User
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"